remove precomputed models list
in the previous version classifier models were reconstructed based on fitted or loaded parameters either when inference finished or when precomputed models were loaded. in posterior evaluation the models  were then passed to the likelihood ratio estimation function in classifier. now the fitted or loaded classifier parameters are passed instead, and the classifier reconstructs the classifier model if needed.