LFIRE classifiers update (#13)
* add BONFIRE classifiers

* add posterior evaluation based on precomputed models

1. added functionalities to save/load models

2. added posterior evaluation based on new observations as a separate function

* integrate precomputed models

aim: use the same functions to initialise LFIRE and calculate posterior probabilities both when precomputed models are available and not

mods:

1. added initialisation based on precomputed models as option in init

2. inference uses precomputed models when available

3. possible to provide new observations as input to inference

* add example notebooks

* rename precomputed models input parameter

* substitute assertions with other python errors

* remove precomputed models list

in the previous version classifier models were reconstructed based on fitted or loaded parameters either when inference finished or when precomputed models were loaded. in posterior evaluation the models  were then passed to the likelihood ratio estimation function in classifier. now the fitted or loaded classifier parameters are passed instead, and the classifier reconstructs the classifier model if needed.

* update docstrings

* add precomputed models save/load test

* store classifiers in the classifier class

in the previous version classifier parameters were stored in LFIRE state

* add example notebook with gaussian process classifier